appease sonar duplication
